pyproject.toml
[project]
name = "webapi"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "django-ninja>=1.5.3",
    "gunicorn>=24.1.1",
    "psycopg[binary]>=3.3.2",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
    "django-redis>=6.0.0",
    "redis>=7.1.0",
    "channels[daphne]>=4.3.2",
    "channels-redis>=4.3.0",
    "openai>=2.24.0",
]
